+++ hakaiApi/NEWS.md
@@ -1,7 +1,9 @@
 # hakaiApi 1.0.3.9000
 
 Enhancements
+
 * the crendentials path is now configurable and environment variables can also be used for tokens (#25)
+* Can now use relative urls once client has been initialized (#27)
 
 Bug fixes
 * use redacted headers so that they can't be accidentally saved to disk (thx @hadley in #28)